add lockdown to deletePost

+++ src/resolvers/Mutation.js
@@ -78,7 +78,20 @@ const Mutation = {
             },
         }, info);
     },
-    deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request);
+        
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!postExists) {
+            throw new Error('unable to delete post');
+        }
+
         return prisma.mutation.deletePost({
             where: {
                 id: args.id
